feat: Add GPU support for CUDA programming
This commit adds GPU support for the code running service. It introduces a property 'NeedGpu' to define if a language needs GPU support or not. The 'RunCode' method in 'RunCodeService' has been modified to check 'NeedGpu' and execute the command accordingly. This change enables support for CUDA programming with Nvidia GPUs.

See merge request aiursoft/cpprunner!39