feat: Add Python with PyTorch language support
This commit adds support for the Python with PyTorch language in the CppRunner project. It introduces a new class `PythonWithPytorch` that implements the `ILang` interface. The class defines the display name, file extension, language name, default code, entry file name, Docker image, run command, and other files for Python with PyTorch. Additionally, the `NeedGpu` property is set to `true` to indicate that this language requires GPU support. This change enables users to write and execute code in Python with PyTorch using the CppRunner service.

See merge request aiursoft/cpprunner!39