feat: Refactor RunCodeService to remove unnecessary parameter
The code change refactors the RunCodeService class in the CppRunner project to remove the 'withGpu' parameter from the RunCode method. The 'withGpu' parameter was previously used to determine if the code should be run with GPU support. However, this information is now available in the ILang interface implementation, so the parameter is no longer needed. This change simplifies the code and improves maintainability.

See merge request aiursoft/cpprunner!39