feat: Update language selection to use language extensions
This commit updates the App component in the Aiursoft.CppRunner.Frontend project to use language extensions instead of language names for the default code and state. Previously, the code used the 'langName' property to set the default code and state language, but now it uses the 'langExtension' property. This change ensures consistency with the recent refactoring of language file extensions to match language names.

See merge request aiursoft/cpprunner!39